+++ kobin/__init__.py
@@ -1,3 +1,4 @@
 from .app import Kobin, Config, current_app, current_config
-from .environs import request, Response, TemplateResponse, JSONResponse, HTTPError
-from .routes import redirect
+from .environs import (
+    request, Response, TemplateResponse, JSONResponse, RedirectResponse, HTTPError
+)